requirements.txt
absl-py==1.2.0
anykeystore==0.2
brotlipy==0.7.0
cachetools==5.2.0
certifi
cffi
charset-normalizer
cryptacular==1.6.2
cryptography
cycler==0.11.0
defusedxml==0.7.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
greenlet==1.1.3
grpcio==1.48.1
hupper==1.10.3
idna
importlib-metadata==4.12.0
joblib==1.2.0
kiwisolver==1.3.2
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
networkx==2.6.3
numpy
oauthlib==3.2.1
opencv-python==4.5.3.56
packaging==21.3
PasteDeploy==2.1.1
pbkdf2
Pillow==9.2.0
plaster==1.0
plaster-pastedeploy==0.7
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pyOpenSSL
pyparsing==3.0.9
pyramid==2.0
pyramid-mailer==0.15.1
PySocks
python3-openid==3.2.0
PyWavelets==1.2.0
repoze.sendmail==4.4.1
requests
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six
SQLAlchemy==1.4.41
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
transaction==3.0.1
translationstring==1.4
typing_extensions
urllib3
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.7
Werkzeug==2.2.2
WTForms==3.0.1
wtforms-recaptcha==0.3.2
zipp==3.8.1
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
